+++ cards/android/src/main/kotlin/com/moengage/flutter/cards/Constants.kt
@@ -18,6 +18,6 @@ internal const val METHOD_IS_ALL_CATEGORY_ENABLED = "isAllCategoryEnabled"
 internal const val METHOD_NEW_CARDS_COUNT = "getNewCardsCount"
 internal const val METHOD_UN_CLICKED_CARDS_COUNT = "unClickedCardsCount"
 
-const val METHOD_INBOX_OPEN_CARDS_SYNC = "onInboxOpenCardsSync"
-const val METHOD_PULL_TO_REFRESH_CARDS_SYNC = "onPullToRefreshCardsSync"
-const val METHOD_APP_OPEN_CARDS_SYNC = "onAppOpenCardsSync"
+internal const val METHOD_INBOX_OPEN_CARDS_SYNC = "onInboxOpenCardsSync"
+internal const val METHOD_PULL_TO_REFRESH_CARDS_SYNC = "onPullToRefreshCardsSync"
+internal const val METHOD_APP_OPEN_CARDS_SYNC = "onAppOpenCardsSync"
