+++ cards/lib/src/internal/constants.dart
@@ -1,4 +1,4 @@
-const String moduleTag = "Cards_";
+const MODULE_TAG = "Cards_";
 
 //Cards Method Channel
 const String cardsMethodChannel = "com.moengage/cards";
