+++ cards/lib/src/model/card.dart
@@ -1,4 +1,4 @@
-import 'package:moengage_cards/src/internal/contants.dart';
+import 'package:moengage_cards/src/internal/constants.dart';
 import 'package:moengage_cards/src/model/template.dart';
 
 import 'meta_data.dart' as moe;
