+++ cards/lib/src/model/image_style.dart
@@ -1,4 +1,4 @@
-import 'package:moengage_cards/src/internal/contants.dart';
+import 'package:moengage_cards/src/internal/constants.dart';
 import 'package:moengage_cards/src/model/widget_style.dart';
 import 'package:moengage_cards/src/model/widget_type.dart';
