+++ cards/lib/src/model/navigation_acion.dart
@@ -1,4 +1,4 @@
-import 'package:moengage_cards/src/internal/contants.dart';
+import 'package:moengage_cards/src/internal/constants.dart';
 import 'package:moengage_cards/src/model/action.dart';
 import 'package:moengage_cards/src/model/navigation_type.dart';
