+++ cards/test/cards_mock_platform.dart
@@ -1,6 +1,6 @@
-import 'package:moengage_cards/src/internal/cards_platform.dart';
+import 'package:moengage_cards/src/internal/cards_controller_android.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 /// Mock Platform Interface.
-class MockCardsPlatform extends MoEngageCardsPlatform
+class MockCardsPlatform extends MoEAndroidCardsController
     with MockPlatformInterfaceMixin {}
